LPS-200485 Move ZipWriterImpl and ZipReaderImpl to module
